refactor(reducer): extract WebPane reduce-block
Move the 18 top-level web-pane Action cases (openWebPanePath,
webPaneFocusURLBar, webPaneOpenNewTab, webPaneTabCycleFocused,
webPaneTabCloseActiveFocused, setWebInspectArmedSubmit,
webBatchInspectStart, webBatchInspectToggle, webBatchInspectHide,
webBatchInspectShow, webPaneSetPrivate, webBatchInspectCancel,
webBatchInspectSend, webBatchFocusItem, syncBatchMarkers,
pushBatchCommentToPage, webBatchDismissPopover, webInspectPayloadReceived)
plus all handleWeb* helpers (and WebPaneScope / WebNavAction /
TabRefResolution / resolveWebPane / runWebPaneActuator /
handleWebActuatorCall / resolveTabRef / tabJSON / consoleLineJSON /
inspectResultJSON / cookieJSON) out of AppReducer.swift into a new
AppReducer+WebPane.swift reduce-block over AppReducer.State.

domain(of:) gains a .webPane arm (still exhaustive, no default).
webPaneReducer is wired into body right after searchNotifyReducer.

The socket web verbs stay in core's .socketMessage handler (Stage 6)
and still call the relocated handleWeb* methods via self. The shared
helpers (resolvePaneTarget / paneSendText / workspaceContainingPane)
and the webPaneURLFocusTokens / webInspectArmedSubmit state fields
stay in core. Behavior-preserving: 1127 tests pass.
